`TestCase`と`Snapshots`をテストごとにファイル分離
